move `compute vocabulary` to its own directory.
